add domUpdates.js file and move all DOM  manipulation from scripts file to it
